#520 Split batch and incremental implementation of MetastoreReader.
