+++ pramen/core/src/main/scala/za/co/absa/pramen/core/metastore/MetastoreReaderBatchImpl.scala
@@ -1,0 +1,39 @@
+/*
+ * Copyright 2022 ABSA Group Limited
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package za.co.absa.pramen.core.metastore
+
+import org.apache.spark.sql.DataFrame
+import org.slf4j.LoggerFactory
+import za.co.absa.pramen.api.MetadataManager
+import za.co.absa.pramen.api.status.TaskRunReason
+import za.co.absa.pramen.core.bookkeeper.Bookkeeper
+
+import java.time.LocalDate
+
+class MetastoreReaderBatchImpl(metastore: Metastore,
+                               metadataManager: MetadataManager,
+                               bookkeeper: Bookkeeper,
+                               tables: Seq[String],
+                               infoDate: LocalDate,
+                               runReason: TaskRunReason) extends MetastoreReaderBase(metastore, metadataManager, bookkeeper, tables, infoDate, runReason) {
+  private val log = LoggerFactory.getLogger(this.getClass)
+
+  override def getCurrentBatch(tableName: String): DataFrame = {
+    log.info(s"Getting daily data for table '$tableName' at '$infoDate'...")
+    metastore.getTable(tableName, Option(infoDate), Option(infoDate))
+  }
+}
